Data geeneration step now saves all data into h5 files from HSDS

+++ molmimic/parsers/container.py
@@ -1,12 +1,19 @@
 import os
 import sys
+import json
+import time
+import types
 import subprocess
 import tempfile
 import shutil
+import warnings
+
 from datetime import datetime
 from io import StringIO
 from collections import OrderedDict
 from subprocess import CalledProcessError
+from contextlib import contextmanager
+
 
 from toil.realtimeLogger import RealtimeLogger
 from toil.job import Job
@@ -23,28 +30,58 @@ CONTAINER_PATH = os.environ.get("CONTAINER_PATH", os.environ["HOME"])
 os.environ["ALLOWABLE_CONTAINER_PATHS"] = "/project"
 
 FORCE_LOCAL = os.environ.get("FORCE_LOCAL", "false")[0].lower()=="t"
-USE_SINGULARITY = os.environ.get("USE_SINGULARITY", "true")[0].lower()=="t"
+USE_SINGULARITY = os.environ.get("USE_SINGULARITY", "false")[0].lower()=="t"
 USE_DOCKER = os.environ.get("USE_DOCKER", "false")[0].lower()=="t"
+
+class ContainerSystemError(RuntimeError):
+    pass
+
+def check_container_system(system):
+    assert system.lower() in ["docker", "singularity"]
+
+    try:
+        rc = subprocess.check_output(["which", system.lower()])
+    except (subprocess.CalledProcessError, FileNotFoundError):
+        raise ContainerSystemError
 
 if FORCE_LOCAL:
     USE_SINGULARITY = False
     USE_DOCKER = False
+    warnings.warn("Using LOCAL for container runner. This feature has not been thoroughly tested -- make sure the correct tool is installed in ordered to run. Please install singualrity or docker to get intended use.")
 elif USE_SINGULARITY:
+    try:
+        check_container_system("singularity")
+    except ContainerSystemError:
+        raise ContainerSystemError("Specified system singularity not found")
+    USE_SINGULARITY = True
     USE_DOCKER = False
-    try:
-        rc = subprocess.check_output(["which", "singularity"])
-        USE_SINGULARITY = True #rc.startswith("usage singularity")
-    except (subprocess.CalledProcessError, FileNotFoundError):
-        USE_SINGULARITY = False
-    FORCE_LOCAL = not USE_SINGULARITY
+    FORCE_LOCAL = False
 elif USE_DOCKER:
-    USE_SINGULARITY = False
     try:
-        rc = subprocess.check_output(["which", "docker"])
-        USE_DOCKER = True #rc.strip().startswith("Usage:	docker")
-    except (subprocess.CalledProcessError, FileNotFoundError):
-        USE_DOCKER = False
-    FORCE_LOCAL = not USE_DOCKER
+        check_container_system("docker")
+    except ContainerSystemError:
+        raise ContainerSystemError("Specified system docker not found")
+    USE_SINGULARITY = False
+    USE_DOCKER = True
+    FORCE_LOCAL = False
+else:
+    #Nothing specified, try docker then singulairty then local
+    try:
+        check_container_system("docker")
+        USE_SINGULARITY = False
+        USE_DOCKER = True
+        FORCE_LOCAL = False
+    except ContainerSystemError:
+        try:
+            check_container_system("singularity")
+            USE_SINGULARITY = True
+            USE_DOCKER = False
+            FORCE_LOCAL = False
+        except ContainerSystemError:
+            USE_SINGULARITY = False
+            USE_DOCKER = False
+            FORCE_LOCAL = True
+            warnings.warn("Using LOCAL for container runner. This feature has not been thoroughly tested -- make sure the correct tool is installed in ordered to run. Please install singualrity or docker to get intended use.")
 
 if USE_SINGULARITY:
     from molmimic.parsers.singularity import apiSingularityCall as containerCall
@@ -64,6 +101,18 @@ if USE_DOCKER or FORCE_LOCAL:
 
 class StoreTrueValue(object):
     pass
+
+def iterator_to_list(func):
+    def wrapper(*args, **kwds):
+        result = func(*args, **kwds)
+        if isinstance(result, types.GeneratorType):
+            result = list(result)
+            if len(result) == 1:
+                return result[0]
+            return result
+        return result
+    return wrapper
+
 
 class Container(object):
     """
@@ -90,11 +139,14 @@ class Container(object):
     RULES = {}
     ENTRYPOINT = None
     RETURN_FILES = False
+    DETACH = False
     ARG_START = ""
     ARG_SEP = " "
+    GPUS = False
+    EXTRA_CONTAINER_KWDS = {}
 
     def __init__(self, job=None, return_files=False, force_local=False, fallback_local=False,
-      intermediate_file_store=None, work_dir=None, detach=False):
+      intermediate_file_store=None, work_dir=None, detach=False, cleanup_when_done=True):
         assert (self.IMAGE, self.LOCAL).count(None) <= 1, "Must define container or local path"
 
         if self.LOCAL is not None:
@@ -107,7 +159,8 @@ class Container(object):
         self.force_local = (FORCE_LOCAL or force_local or self.IMAGE is None) and self.LOCAL is not None
         self.fallback_local = fallback_local or self.LOCAL is not None
         self.return_files = return_files or self.RETURN_FILES
-        self.detach = detach
+        self.detach = detach or self.DETACH
+        self.cleanup_when_done = cleanup_when_done
         self.job = job
         self.rules = {
             str: lambda k, v: str(v),
@@ -129,6 +182,20 @@ class Container(object):
                 datetime.now().strftime('%Y-%m-%d-%H:%M:%S')
             ))
 
+        self.process_args()
+
+        if self.GPUS:
+            self.enable_gpus()
+
+        # if not self.detach:
+        #     self.__call__ = iterator_to_list(self.__call__)
+
+    def __init_subclass__(cls, *args, **kwargs):
+         super().__init_subclass__(*args, **kwargs)
+         if not cls.DETACH:
+             cls.__call__ = iterator_to_list(cls.__call__)
+
+    def process_args(self):
         self.parameters = []
         self.param_names = []
         self.param_funcs = {}
@@ -236,11 +303,19 @@ class Container(object):
 
         parameters = self.format_parameters(args, kwds)
         RealtimeLogger.info(parameters)
+        print(parameters)
 
-        image = pullContainer(self.IMAGE, pull_folder=CONTAINER_PATH)
+        image = self.IMAGE
+        if USE_DOCKER:
+            image = image.replace("docker://", "")
+
+        image = pullContainer(image, pull_folder=CONTAINER_PATH)
+
+        if USE_SINGULARITY:
+            self.EXTRA_CONTAINER_KWDS["return_result"] = True
 
         try:
-            with silence_stdout(), silence_stderr():
+            if True: #with silence_stdout(), silence_stderr():
                 out = containerCall(
                     self.job,
                     image=image,
@@ -248,32 +323,48 @@ class Container(object):
                     working_dir="/data",
                     volumes={self.work_dir:{"bind":"/data", "mode":"rw"}},
                     parameters=parameters,
-                    return_result=True,
-                    detach=self.detach)
+                    detach=self.detach,
+                    **self.EXTRA_CONTAINER_KWDS)
         except (SystemExit, KeyboardInterrupt):
             raise
-        except:
-            if self.fallback_local:
-                import traceback as tb
-                RealtimeLogger.error(tb.format_exc())
-                return self.local(*args, **kwds)
-            self.clean()
-            self.change_paths = OrderedDict()
-            raise
+        except Exception as e:
+            if "chown: changing ownership" in str(e):
+                #Handle bug when running on samba shares
+                raise
+            else:
+                if self.fallback_local:
+                    import traceback as tb
+                    RealtimeLogger.error(tb.format_exc())
+                    return self.local(*args, **kwds)
+                self.clean()
+                self.change_paths = OrderedDict()
+                if self.GPUS and USE_DOCKER:
+                    print("In order to use GPUs with docker, please install nvidia docker container toolkkit: https://docs.nvidia.com/datacenter/cloud-native/container-toolkit/install-guide.html#docker")
+                raise
 
         if self.detach:
             message = ""
-
             try:
-                for line in out:
-                    message += line
+                if USE_DOCKER:
+                    out = out.logs(stream=True)
+                    for line in out:
+                        line = line.decode("utf-8")
+                        yield line
+                else:
+                    for line in out:
+                        yield line
+
+                    #yield from out
+
             except (SystemExit, KeyboardInterrupt):
                 raise
-            except:
+            except Exception as e:
+                raise
                 #Singularity raises subprocess.CalledProcessError, but not sure about Docker
-                self.clean()
+                #self.clean()
                 self.change_paths = OrderedDict()
-                raise RuntimeError(message)
+                import traceback as tb
+                raise RuntimeError(f"Output: {message}, Error: {tb.format_exc()}")
 
         else:
             if USE_SINGULARITY:
@@ -282,7 +373,7 @@ class Container(object):
                 if out["return_code"]:
                     self.clean()
                     self.change_paths = OrderedDict()
-                    raise RuntimeError(message)
+                    raise RuntimeError(f"{parameters} {out}")
             else:
                 #Docker already handled error above
                 message = out
@@ -295,14 +386,21 @@ class Container(object):
             self.change_paths = OrderedDict()
             raise
 
+        self.out_files = out_files
+
         if self.return_files:
             self.stdout = message
             self.clean()
             self.change_paths = OrderedDict()
-            return out_files
 
-        self.change_paths = OrderedDict()
-        return message
+            if not self.detach:
+                yield out_files
+        else:
+            self.change_paths = OrderedDict()
+
+            self.message = message
+            if not self.detach:
+                yield message
 
     def local(self, *args, **kwds):
         self.is_local = True
@@ -332,6 +430,10 @@ class Container(object):
         if len(args)+len(kwds) == self.number_of_parameters:
             pass
 
+
+        RealtimeLogger.info(f"args {args}")
+        RealtimeLogger.info(f"kwds {kwds}")
+
         # if len(args) == self.number_of_parameters and len(kwds) == self.number_of_optional_parameters":
         #     #Correct number of args, all default options
         #     kwds = zip(self.param_names, args).update(kwds)
@@ -346,13 +448,17 @@ class Container(object):
             params = dict(zip(self.param_names[:len(args)], args))
             need_params = self.param_names[len(args):]
             if len(need_params) == 0 or all(k in kwds for k in need_params):
-                kwds = params.update(kwds)
+                params.update(kwds)
+                kwds = params
             else:
                 raise RuntimeError
 
         parameters = [p for p in self.parameters] #[[p] for p in self.parameters]
 
+        RealtimeLogger.info(f"p {parameters}")
+
         for k, v in kwds.items():
+            RealtimeLogger.info(f"checking {k}={v}")
             try:
                 idx = self.params_to_update[k]
                 param_func = self.param_funcs[idx]
@@ -391,12 +497,26 @@ class Container(object):
 
     def arg_formatter(self, key, value, formatter):
         if isinstance(formatter, (list, tuple)):
+            #If value is list, choose the startinf arg
+            start_arg = 0
+
             #Assume correct formatting
             args = []
             for part in formatter:
-                if formatter.count("{}") == 1:
+                if isinstance(value, (list, tuple)):
+                    if len(value) <= len(formatter):
+                        if part.count("{}") == 1:
+                            args.append(part.format(value[start_arg]))
+                            start_arg += 1
+                        elif part.count("{}") == 0:
+                            args.append(part)
+                        else:
+                            raise RuntimeError("Value must contain only one formatter string")
+                    else:
+                        raise RuntimeError("Value list must be <= to formatter list")
+                elif part.count("{}") == 1:
                     args.append(part.format(value))
-                elif formatter.count("{}") == 2:
+                elif part.count("{}") == 2:
                     args.append(part.format(key, value))
                 else:
                     args.append(part.format(key=key, value=value))
@@ -422,16 +542,34 @@ class Container(object):
             raise RuntimeError(f"Invalid arg formatter: {formatter}")
 
     def format_in_path(self, name, path, move_files_to_work_dir=True):
-        if self.is_local or not move_files_to_work_dir or any(path.startswith(p) for p in os.environ.get("ALLOWABLE_CONTAINER_PATHS", "").split(":")):
+        if False and self.is_local or not move_files_to_work_dir or any(path.startswith(p) for p in os.environ.get("ALLOWABLE_CONTAINER_PATHS", "").split(":")):
             if not os.path.isfile(path):
                 raise RuntimeError("{} is not found".format(path))
             return os.path.abspath(path)
         else:
-            new_path = os.path.join("/data", os.path.basename(path))
-            local_path = os.path.abspath(os.path.join(self.work_dir, os.path.basename(path)))
             if not os.path.abspath(os.path.dirname(path)) == os.path.abspath(self.work_dir):
-                shutil.copy(path, local_path)
-                self.files_to_remove.append(local_path)
+                if str(os.path.abspath(path)).startswith(os.path.abspath(self.work_dir)):
+                    start_path = path
+                    while os.path.abspath(start_path) != os.path.abspath(self.work_dir):
+                        start_path = os.path.dirname(start_path)
+
+                    path = path[len(start_path):]
+                    if path.startswith("/"):
+                        path = path[1:]
+
+                    new_path = os.path.join("/data", path) if not self.is_local else path
+                else:
+                    local_path = os.path.abspath(os.path.join(self.work_dir, os.path.basename(path)))
+                    if not os.path.isfile(local_path):
+                        shutil.copyfile(path, local_path)
+                        #self.files_to_remove.append(local_path)
+                    new_path = os.path.join("/data", os.path.basename(path)) if not self.is_local else os.path.basename(path)
+                    if not os.path.isfile(local_path):
+                        time.sleep(2)
+                    assert os.path.isfile(local_path), f"DNE {local_path}"
+            else:
+                new_path = os.path.join("/data", os.path.basename(path)) if not self.is_local else os.path.basename(path)
+
             return new_path
 
     def format_out_path(self, name, path):
@@ -491,21 +629,62 @@ class Container(object):
         return out_files
 
     def clean(self):
+        if not self.cleanup_when_done:
+            return
         while len(self.files_to_remove) > 0:
             try:
                 os.remove(self.files_to_remove.pop())
             except OSError:
                 pass
 
+    def enable_gpus(self, gpu=None):
+        if USE_SINGULARITY:
+            self.EXTRA_CONTAINER_KWDS["nv"] = True
+            os.environ["SINGULARITYENV_CUDA_VISIBLE_DEVICES"] = str(gpu)
+        elif USE_DOCKER:
+            self.EXTRA_CONTAINER_KWDS["runtime"] = "nvidia"
+            self.EXTRA_CONTAINER_KWDS["environment"] = [f"CUDA_VISIBLE_DEVICES={gpu}"]
+
+    def disable_gpus(self):
+        if USE_SINGULARITY:
+            if "nv" in self.EXTRA_CONTAINER_KWDS:
+                del self.EXTRA_CONTAINER_KWDS["nv"]
+        elif USE_DOCKER:
+            if "runtime" in self.EXTRA_CONTAINER_KWDS:
+                del self.EXTRA_CONTAINER_KWDS["runtime"]
+
+    @contextmanager
+    def custom_entrypoint(self, entry_point, args):
+        old_entrypoint = self.ENTRYPOINT
+        old_local = self.LOCAL
+        old_args = self.PARAMETERS
+        self.ENTRYPOINT = self.LOCAL = entry_point
+        self.PARAMETERS = args
+        self.process_args()
+        yield
+        self.ENTRYPOINT = old_entrypoint
+        self.LOCAL = old_local
+        self.PARAMETERS = old_args
+        self.process_args()
+
+    @contextmanager
+    def custom_parameters(self, args):
+        old_args = self.PARAMETERS
+        self.PARAMETERS = args
+        self.process_args()
+        yield
+        self.PARAMETERS = old_args
+        self.process_args()
+
     @classmethod
-    def tempfile(cls, work_dir=None):
+    def tempfile(cls, delete=False, work_dir=None):
         if work_dir is None:
             if hasattr(cls, "work_dir"):
                 work_dir = cls.work_dir
             else:
                 work_dir = os.getcwd()
 
-        tfile = tempfile.NamedTemporaryFile(dir=work_dir, delete=False)
+        tfile = tempfile.NamedTemporaryFile(dir=work_dir, delete=delete)
         tfile.close()
         return tfile.name
